Rename components/parser to tags/dispatch
