+++ include/events/signal.hpp
@@ -2,8 +2,8 @@
 
 #include "common.hpp"
 #include "components/ipc.hpp"
-#include "components/parser.hpp"
 #include "components/types.hpp"
+#include "tags/dispatch.hpp"
 #include "tags/types.hpp"
 #include "utils/functional.hpp"
